+++ __project__/settings.py
@@ -54,6 +54,7 @@ INSTALLED_APPS = [
 
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
+    'whitenoise.middleware.WhiteNoiseMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
     'django.middleware.locale.LocaleMiddleware',
     'django.middleware.common.CommonMiddleware',
@@ -144,6 +145,11 @@ STATIC_URL = 'static/'
 STATICFILES_DIRS = [
     PROJECT_DIR / "static"
 ]
+
+# Use whitenoise to serve static files
+# See http://whitenoise.evans.io/en/stable/django.html.
+STATICFILES_STORAGE = 'whitenoise.storage.CompressedManifestStaticFilesStorage'
+
 
 # Default primary key field type
 # https://docs.djangoproject.com/en/4.1/ref/settings/#default-auto-field