+++ CHANGELOG.md
@@ -13,3 +13,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Configure project .env file consumption
 - Configure django static files management
 - Configure project internalization machinery
+- Setup whitenoise to serve static files